Seperated EntertainmentConsole class object into its own file.
